Chore/#16 더미 데이터 추가 및 스키마 수정 (#18)
* chore(data.sql): data.sql 추가 [#16]

더미데이터 추가를 위한 data.sql 작성

* chore(schema.sql): 테이블 수정 [#16]
Discount 테이블 amount BIGINT -> FLOAT
Users 테이블 pay_password NOT NULL -> NULL
Benefit 테이블 merchant_id NOT NULL -> NULL

* chore(sql): dummy data 입력문 수정 및 혜택 추가 [#16]
현대카드 M의 모든 혜택을 추가하였습니다.

* chore(sql): schema.sql 수정 [#16]
benefit_condition 테이블의 category 컬럼 enum type 수정

* chore(sql): data.sql 최종 수정 [#16]

직접 데이터베이스에 INSERT 시에 발생하는 오류들을 해결하고 더미 데이터가 입력된 것을 확인하였습니다.

* chore(sql): 최종 스키마 수정 [#16]
실제 데이터에 맞게 스키마를 최종 수정하였습니다.
수정 사항들은 PR에 상세하게 적어두겠습니다.

* chore(sql): 본인 인증 키 추가 [#16]

본인 인증 키를 USER TABLE에 추가하였습니다.

* chore(JPA): 스키마 변경에 따른 JPA 수정 [#16]
변경된 스키마에 맞춰 JPA Entity 수정
Nullable 제거
personalAuthKey 추가

+++ src/main/java/com/tower_of_fisa/paydeuk_server_service/domain/entity/Discount.java
@@ -6,7 +6,6 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @Entity
@@ -23,7 +22,7 @@ public class Discount extends BaseEntity {
     private DiscountApplyType applyType;
 
     @Column(name = "amount", nullable = false)
-    private Long amount;
+    private Float amount;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "benefit_id", nullable = false)
